requirements.txt
astroid==3.0.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
dill==0.3.7
Flask==3.0.0
idna==3.6
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mccabe==0.7.0
Naked==0.1.32
packaging==23.2
platformdirs==4.0.0
pluggy==1.3.0
pycparser==2.21
PyJWT==2.8.0
pylint==3.0.2
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.8.0
shellescape==3.8.1
six==1.16.0
tomlkit==0.12.3
urllib3==2.0.7
Werkzeug==3.0.1
